IMutableTransform.ToImmutableを実装
