+++ src/BeUtl.Graphics/Graphics/Effects/IMutableBitmapEffect.cs
@@ -1,0 +1,8 @@
+﻿using BeUtl.Media;
+
+namespace BeUtl.Graphics.Effects;
+
+// Todo: IAnimatable
+public interface IMutableBitmapEffect : IBitmapEffect, ICoreObject, IAffectsRender
+{
+}
