+++ src/BeUtl.Graphics/Graphics/Filters/IImageFilter.cs
@@ -1,0 +1,12 @@
+﻿using SkiaSharp;
+
+namespace BeUtl.Graphics.Filters;
+
+public interface IImageFilter
+{
+    bool IsEnabled { get; }
+
+    Rect TransformBounds(Rect rect);
+
+    SKImageFilter ToSKImageFilter();
+}
