+++ src/BeUtl.Graphics/Graphics/Filters/IMutableImageFilter.cs
@@ -1,0 +1,8 @@
+﻿using BeUtl.Media;
+
+namespace BeUtl.Graphics.Filters;
+
+// Todo: IAnimatable
+public interface IMutableImageFilter : IImageFilter, ICoreObject, IAffectsRender
+{
+}
