+++ src/BeUtl.Graphics/Graphics/Transformation/IMutableTransform.cs
@@ -1,0 +1,9 @@
+﻿using BeUtl.Media;
+
+namespace BeUtl.Graphics.Transformation;
+
+// Todo: IAnimatable
+public interface IMutableTransform : ICoreObject, ITransform, IAffectsRender
+{
+    ITransform ToImmutable();
+}
