+++ src/BeUtl.Graphics/Graphics/Transformation/ITransform.cs
@@ -1,15 +1,8 @@
-﻿using BeUtl.Media;
-
-namespace BeUtl.Graphics.Transformation;
+﻿namespace BeUtl.Graphics.Transformation;
 
 public interface ITransform
 {
     bool IsEnabled { get; }
 
     Matrix Value { get; }
-}
-
-// Todo: IAnimatable
-public interface IMutableTransform : ICoreObject, ITransform, IAffectsRender
-{
 }
